fix some bug

+++ validation/public.ts
@@ -1,6 +1,6 @@
 import { number, object } from "zod";
 
 export const publicPlaylistData = object({
-    accountId: number(),
-    playlistId: number()
+    accountId: number().int(),
+    playlistId: number().int()
 })
